DrawBoolean ported
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@15012 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/awt/Dimension.java
@@ -2,7 +2,7 @@ package geogebra.common.awt;
 
 public abstract class Dimension {
 
-	public abstract double getWidth();
-	public abstract double getHeight();
+	public abstract int getWidth();
+	public abstract int getHeight();
 
 }
